Create BusinessProcessSpecification container for annotations
Up to this commit, the new classes for business process specifications have been directly referenced from the annotation repository (1:1 containment reference). They are part of a container class BusinessProcessSpecification now; only the container is directly linked to the annotation repository and is able to hold any number of each business process annotation type (similar to other annotations type which are also contained in "XyZSpecification" containers).