Add parent relationship for BP annotations, updated genmodel
When moving the business process annotation in a new container class BusinessProcessSpecification, the containment references from the items to the Specification have been accidentally changed to a simple reference. This is resolved. Furthermore the genmodel is updated so all the new annotations are available when starting an inner Eclipse (without the update, only the BPSpecification element could be used).