Create projects for bp-related modificationmarks
Similar to the separation of field of activity annotations between the old KAMP and the new approach for business processes, the business process-related modificationmarks are also placed in new projects (the new model only contains a repository class at the moment, it will be extended in the future). The new projects are not referenced by the kampintbiis main project yet.