BP modifications I
The modificationmarks model is extended with new modification types. This commit contains the types that have been added to the original KAMP, as they do not refer to business process-specific classes, but could also be used without those extensions. EClasses for branch, loop and entry level system call modifications are part of the extended model.