Save auto-generated changes in modificationmarks
Some files are changed after generating the model without any changes in the model. This is due to implements-declaration changes or whitespace issues. This commit summarizes those changes for the modificationmarks project. Not committed is the modificationmarksEditor file, as it contains manual changes which are not preserved with "generated NOT". Such annotations would have to be added before a new generation.