Update KAMP modificationmarks to newer EMF
When auto-generating the model+editor code for modificationmarks, some changes have been added only because of the different EMF versions between the creation of KAMP and now (e.g. differences in property sheets, diagnostics). This commit summarizes those changes in modificationmarksEditor (some non-critical modifications in the projects have already been part of the recent commit 29631). Furthermore the manually added code in the modificationmarksEditor has been restored, the relevant methods marked with "generated Not" and the affected code sections highlighted with comments around them.