Update IntBIIS (bppcm) to newer EMF
When auto-generating the model+editor code for IntBIIS, some changes have been added only because of the different EMF versions between the creation of IntBIIS and now (e.g. differences in property sheets, diagnostics). During the migration of IntBIIS to the latest Eclipse, those generation changes have been reverted before committing (because of uncertainty if the older code has been manually added and should not be changed), but now the code in the repository should equal the state after a full generation . Furthermore the manually added code in the ..ResourceFactoryImpls has been restored, the relevant methods marked with "generated Not" and the affected code sections highlighted with comments around them.