Change class relationship of BPModificationRepository
The new modification repository with business process features has been a subclass of the original ModificationRepository without adding further attributes. This solution is problematic concerning the types, as the subclass repository should also have a subclass attribute BpSeedModifications instead of the superclass SeedModifications (so we would need covariance concerning the parameter). When the BpModificationRepository just inherits the attributes, the graphical editor only offers the possibility to add those modifications described by the original ModificationRepository, not the bp-specific ones. Two solutions have been discussed:
1) Making the type of the attribute generic, e.g with a AbstractModificationRepositors<T extends SeedModification> and its subclasses ModificationRepository/BPModificationRepository inserting the types.
2) Destroy the inheritance relationship between ModificationRepository and BPModificationRepository, just duplicate the attributes and change the type. This destroys substitutability which is not needed at the moment.

The second solution is implemented with the current commit.