Change class relationship of BPModificationRepository II
The recently chosen approach of the BPModRepo-ModRepo relationship has proven a bad idea concerning non-generated classes, where the modification repository is accessed at multiple locations for calculation propagation, deriving the workplan, storing the architecture etc. To extend the old code with minimum effort, a kind of inheritance between BpModRepo and ModRepo is necessary. Therefore an abstract, generic super class AbstractModificationRepository has been created in the model (with a parameter T extends SeedModifications) and ModificationRepository as well as BPModificationRepository are both subclasses with a fixed type parameter (and no other attributes).