Adapt "Create empty base model" popup
KAMP defines three different actions that are displayed in the inner workspace when right-clicking on projects. On of them is the creation of new models with all necessary files (.system, .repository etc.). This features is available in an adapted form for KAMP+IntBIIS now (the generated .fieldofactivityannotations and .modificationmarks files will include all possibilities of KAMP+IntBIIS.