Change Bpmodificationmarks model assistant
The assistant for creating KAMP+IntBIIS modificationmarks files automatically selects the BPModificationRepository node as root of the tree and adds the BPSeedModifications node. This pre-configuration is quite similar to the .fieldofactivityannotations assistant that also sets up a repository as root and adds some children. This improves the usability, as users do not need to choose an initial node themselves.