Adapt "Derive workplan" popup action
This action is displayed in the inner workspace when right-clicking on projects. The feature is available in an adapted form for KAMP+IntBIIS now. To improve extensibility of difference calculation and workplan derivation (which will be modified for business processes in the future), some static methods have been made non-static - although the objects which are used now to call those methods are in consequence singletons, the use of non-static methods allows inheritance and dynamic dispatch for adding new code. The corresponding subclasses for KAMP+IntBIIS have been created, but apart from inheriting from the original KAMP classes (and initializing their fields in the constructor) they have no further functionality at the moment.