Adapt modificationmarks.editor actions
The major context menu actions defined by the modificationmarks plugin are step 3 (calculate change propagation) and step 4 (calculate activities) of the KAMP algorithm (the user decision regarding the modifications are in a to do-state with empty method bodies in the original KAMP). Both step are available in KAMP+IntBIIS now, the necessary classes inheriting from the corresponding KAMP classes without adding functionality (at the moment).