Extend propagation algorithm with DataType -> EntryLevelSystemCall
This is the first extension of the propagation algorithm for KAMP+IntBIIS, considering the propagation from DataTypes to EntryLevelSystemCalls. Although the approach is quite similar to the DataType->Interface propagation (apart from a different navigation path through the model), a lot of code has been changed to make the KAMP parts of the algorithm extensible. This has involved the introduction of generics to get covariant attributes as well as making static methods non-static to use the benefits of dynamic dispatch. Furthermore the usagemodel is integrated into the persistency (loading from files into objects and the other way around).