Configure IntBIIS/KAMP+IntBIIS model wizards
The model wizards are used when creating a new model via the Eclipse "New" dialogue. Although those classes are generated, some options are hard-coded in four of the wizards now, especially the root element of each model is predefined (in the default code, you get an extra page in the wizard to start with any model element).