Extend propagation algorithm with DataType -> DataObject
Another step is added to the propagation algorithm, marking all dataObjects that have a marked dataType. The code is quite similar to DataType->Interface / DataType->ELSC or even simpler regarding model tree traversal, as the lookup of all existing dataObjects can be done directly with an iterator over the recently created "DataModel" class ("dataModel" attribute of the BPArchitectureVersion).