Make ChangePropagationAction stateless (again)
Before the KAMP+IntBIIS changes, ChangePropagationAction created a new ChangePropagationAnalysis during each run. The creation was moved to the constructor in a refactoring; therefore the object could be stored in a field. This is reverted now, as ChangePropagationAnalysis itself has some fields that are used during the analysis, so we need a fresh object for each change propagation run (we do not want the old results to interfere). As a consequence, the run() method calls a create...() method and gets new instances instead of accessing a field via a getter method (the fields are removed, so ChangePropagationAction is stateless).