Refactor ChangePropagationAnalysis
Another part of the main propagation analysis method (interface to role, inter-component, intra-component) is extracted to a sub-routine, as it is currently the same for KAMP and KAMP+IntBIIS. The creation and the use of the ChangePropagationDueToDataDependencies object is restructured. Especially the relationship ChangePropagationDueToDataDependencies <-> ArchitectureVersion#ModificationRepository#ChangePropagationStep will play an important role when new change propagations are introduced (as DataObject -> DataType), as changes have to be marked not only in the ChangePropagationDueToDataDependencies field, but also in the ArchitectureVersion to prevent duplicates of the same modification or cycles.