Workplan derivation for DataObjects
The DifferenceCalculation and ExcelWriter are adapted to process the addition and removal of DataObjects, integrating those activities into the BPDeriveWorkplanAction (which can be performed by right-clicking on the top-level element of a modificationsmarks file). A lot of other classes have been modified during the process to enable the integration of new elements in existing methods, but only a few classes will have to be changed for the integration of further KAMP+IntBIIS elements in the future. The pure modification marks (without adding or removing) of DataObjects do not appear in the derived Excel file at the moment; the InternalModificationDerivation would have to be changed for this (part of future commits).