Differentiate between base and target version in example
During the adaption of the workplan derivation algorithm for KAMP+IntBIIS, it has become clear that this activity ("Step 4: Calculate activities") requires two versions of the important architecture files, a target version (modified architecture) in one folder and a base version in the parent folder. As a consequence, copies of these files are placed in a new folder called "modified" now.