Enhance datamodel property labels
The DataObjects in the DataModel have an attribute for the DataType. If you use the graphical properties editor of Eclipse, by default the result of the corresponding ItemProvider#getText() from Palladio will be displayed as attribute value, which currently shows the id of the DataType instead of the entityName. By extending the ItemProviderAdapterFactory of the Palladio repository and including that subclass ItemProviderAdapterFactory in the DatamodelEditor, the entityNames are shown now. A clean solution should directly change those ItemProviders in Palladio, but this would affect too many other stakeholders involved in the Palladio project.