Extend propagation algorithm with ELSC -> Interface
Another step is added to the propagation algorithm, considering EntryLevelSystemCall -> Interface propagation besides DataType -> Interface propagation. Duplicates are recognized and eliminated.

Furthermore, the step DataType -> DataObject is placed before DataObject -> DataType. As a consequence, a modification mark for one DataObject will not propagate to all other DataObjects of the same type as before.