Internal modification derivation for DataObjects, ELSC, ResourceActions, Loop, Branch, ActorStep
After some KAMP+IntBIIS-relevant model elements have already been considered for the workplan when adding or removing them in the modified architecture (commit 29888), this commits extends the support to modification marks of those elements. For example, if you mark an ELSC as modified in the *.modificationmarks file, that step will appear in the Excel workplan now.