Add pcm main, pcm.edit, pcm.editor with child creation
One problem when working on KAMP+IntBIIS examples was that IntBIIS sub-classes of abstract PCM classes where not shown as possible elements in the graphical model editor, even if you created an IntBIIS model (e.g. IntBIIS bpusagemodel with classes like ActorStep extending PCM's AbstractUserAction). This commit integrates the original PCM code into the KAMP+IntBIIS repo with one modification: the "childCreationExtenders" attribute of the package "pcm" in the pcm.genmodel file is set to "true".