Rename InterActivityPropagation
The name "InterActivityPropagation" is not accurate for that model element, as it should include various propagations from AbstractUserActions to other AbstractUserActions. Activities are only a special type of AbstractUserAction that encapsulates sub-processes. Therefore the EClass is renamed. It has not been used yet, so the changes only affect generated classes.