Fix AbstractInternalModificationDerivation#deriveSignatureModifications()
The original KAMP code included a lookup for existing signature modification activities. If one was present (!= null), it was immediately replaced. This seems to be a bug, the logic should rather be if == null, create a new. Therefore a fix is made in one line of code. Anyway, that piece of code is not executed at the moment, so the consequences of the fix cannot be seen.