Recursive AbstractUserAction lookup
Currently four different AbstractUserActions are looked up in the model for the propagation algorithm: ELSC, Acquire-, ReleaseDeviceResourceAction and ActorStep. The previous lookup took the path UsageModel -> UsageScenarios -> ScenarioBehaviour -> AbstractUserActions. It did not consider that an AbstractUserAction can be an Activity, Branch or Loop element that has nested ScenarioBehaviours with nested AbstractUserActions. This commit fixes that bug (which was not visible as the current example .usagemodel file does not contains such a special kind of AbstractUserAction) by introducing a recursive lookup.