Change Modify... EClasses to contain Operation... elements
The original KAMP modificationsmarks meta-model was not consistent regarding the EReferences to modified elements. ModifyInterface had a reference to OperationInterface (not to the super class Interface), but ModifySignature referenced Signature instead of OperationSignature, ModifyRequired-/-ProvidedRole referenced Required-/ProvidedRole. As we are currently only interested in the Operation... model elements, the modificationmarks Ecore meta-model contains only references to Operation... elements (where such elements exist) after this commit.