Extend propagation algorithm with Signature -> Signature step
Another step is added to the propagation algorithm, considering propagation from Signatures to other Signatures based on the ComponentInternalDependencies. The old KAMP algorithm only considered role-to-role dependencies, the new version goes one step deeper and also adds modificationmarks for affected signatures on the provided side if at least one signature on the required side of a operation-to-operation dependency (children of the role-to-role dependencies) is also marked.