Change modificationmarks model for DeviceResources
The old model considered the propagation from DeviceResources to the related acquire and release actions. The next goal is to mark all actions between acquire and release, too. Therefore a new meta-model element "DeviceResourcePropagation" is introduced which encapsulates the modifications of the resource (mandatory), acquire action (optional), release action (optional) and all user actions (0..*) between. The changes up to now have nearly no impact on the current propagation algorithm, the ModifyAcquire-/-ReleaseDeviceResourceActions are still treated as ordinary user action modifications and added to the propagation step without using the new class (but this will change in the near future).