Remove some typecasts to Operation... classes
Some casts in the KAMP/KAMP+IntBIIS code have become unnecessary because of the consistent use of OperationInterface, -Signature, -ProvidedRole, -RequiredRole in KAMP's Modificationmarks and ComponentInternalDependencies model since the latest commits. Those explicit casts are removed. There are still some remaining ugly casts in the code, as components (e.g. BasicComponent) are PCM elements (we should not make adaptions there) and inherit from InterfaceProvidingRequiringEntity, which refers to the superclasses ProvidedRole and RequiredRole instead of the Operation... subclasses. Furthermore the interfaces in the PCM repository have the type Interface instead of OperationInterface.