Add warning for erroneous device actions and refactor activity types
A special kind of warning text will be printed to the Excel sheet if a DeviceResourcePropagation contains no acquire action or no release action. To include those changes, the enum ActivityType had to be "extended" (by introducing an interface with a getName() method which is implemented by KAMP's ActivityType as well as KAMP+IntBIIS' new BPActivityType - enum inheritance is not supported by Java). ActivityElementType and BPInternalModficationDerivation have been refactored as well.