Improve ActorStep -> ActorStep step
The recent introduction of the ActorStep -> ActorStep propagation was based on a primitive heuristic: All ActorSteps after marked ActorSteps were also marked, continuing recursively. However, this lead to two weaknesses. Firstly, indirect successors were not considered. They should be considered, as the order might be arbitrary in some cases (there is no element for parallelism) or e.g. a Delay action might be between ActorSteps, not breaking their logical connecting, just representing a time span. Secondly, there was no semantics for the propagation, only the order of steps (direct succession) acted as an indicator for a relationship.
This commit contains a far more advanced solution. ActorSteps are considered to be marked if at least one of their inputDataObjects is an outputDataObject of a direct or indirect predecessor (marked ActorStep). Besides the semantics, the algorithm is still recursive, so newly marked ActorSteps during the propagation step might influence their successors. Branches and Loops are also considered, especially regarding the successor relationship (different parallel branches [in the meta-model called BranchTransitions] do not influence each other, but successors within the BranchTransition and after the Branch element).
The meta-model for modifications is changed such that modifications of ActorSteps can have multiple child elements to reference modifications of preceding ActorSteps (previous version: multiplicity at most 1, because only direct successors considered). The workplan derivation is adapted accordingly.