Refactor change propagation algorithm
The requirements part of the propagation algorithm and model lookup are simplified by method extraction and more compact code. The number of full model traversals is reduced by keeping marked elements in memory (comparable to the business process propagation algorithm). The older algorithms are refactored as well with only some minor changes.