Extend propagation algorithm with Architecture -> Option step
Similar to the propagation of modified Options to DataTypes/Interfaces/Components, this step goes the opposite direction, meaning modified DataType/Interface/Component -> Option. The logic for checking if architecture elements and options are related is the same in this step as in the existing one, but the methods needed to be heavily refactored so they can actually be used in both directions now.