Extend propagation algorithm with Option -> Decision/Requirement step
Similar to the propagation of modified Decisions and Requirements to Options, this step goes the opposite direction, meaning modified Option -> Decision/Requirement. The logic for checking if decisions/requirements and options are related is quite simple and nearly the same in this step as in the opposite direction (checking selectedBy/triggeredBy).