Refactor specification modificationmarks model
The old Toometa modificationmarks meta-model for Decisions, Options and Requirements contained many sub-classes of ChangePropagationStep and was not consistent to the recently added modification elements. This commit standardizes the meta-model. One element ("ChangePropagationDueToSpecificationDependencies") acts as a container for all specification-related modification marks, improving understandability and enabling a more sophisticated workplan generation (based on the improved structure; the old version contained a full model traversal in search of Decision(Option/Requirement modifications). All propagations between Decisions, Requirements and Options can be represented with the "SingleSpecificationPropagation" EClass whose EReferences to the source of the propagation and the affected elements are based on the common super-class (of Decision, Option, Requirement) TraceableObject. This design decision reduces the number of EClasses significantly, but has the drawback that the EReferences of SingleSpecificationPropagation are somewhat weakly-typed (which is not relevant for displaying the modifications, but for the workplan -> simulated polymorphism).
The workplan generation is enhanced for all specification-related modification elements and displays the causing modification as well as the affected elements (old version: only the affected elements).